revno: 1780 message:   Fixed bug, could not delete maplegendset which was member of a mapview. This was caused by a nullpointerexception which happened since the maplegendset property of mapview is optional.
committer: Lars <larshelg@larshelg-laptop>
branch nick: trunk
timestamp: Tue 2010-04-13 16:05:21 +0200

+++ dhis-2/dhis-services/dhis-service-mapping/src/main/java/org/hisp/dhis/mapping/MapViewDeletionHandler.java
@@ -115,7 +115,7 @@ public class MapViewDeletionHandler
     {
         for ( MapView mapView : mappingService.getAllMapViews() )
         {
-            if ( mapView.getMapLegendSet().equals( mapLegendSet ) )
+            if ( mapView.getMapLegendSet() != null && mapView.getMapLegendSet().equals( mapLegendSet ) ) // MapLegendSet is optional
             {
                 mappingService.deleteMapView( mapView );
             }
